Remove duplicate entry in EXPECTED_BLOCKS consts.
The 8th entry was always the same as the 6th, except the Z test.

In C++ the Z test is different from every other test, blocking counterclockwise rotation when other tests block clockwise rotation. I suppose I did this for coverage, but in Rust Z is tested by the same per_block function which expects to block clockwise rotation, and so the data copied from C++ was erroneous in the Rust test.

Ideally I'll go back and have per_block check blocking CCW as well as CW rotation. Possibly make it an argument. We still won't need an 8th entry as blocking CCW means reusing the 4th entry instead of the 6th.