Check MockPiece state
In the C++ this stuff was handled by static state, but Rust restricts statics severely in ways that block straightforward use of std::sync objects. After a few false starts trying to duplicate the static state strategy, I realized that the render callbacks should be passed a reference to the Mock objects that will be keeping the counts I want. In test_run_callback() I used a closure over a mutex to write data from the render function into the test function. Using the same strategy to read out of the MockPiece is a better solution than synchronizing access to static state.

One thing that failed to duplicate the static state strategy was thread-local variables. I expect the relevant functions to be called on internal threads owned by the Game struct, and so writes to thread-locals won't be visible to the test functions' threads.