Callback synch hell
Why are callbacks stored as Arc<Mutex<Box<Fn + Send>>? Where did 'closure go?

The root of the issue is thread::spawn, which demands Send + Sync + 'static.
The Fn has to be Send + 'static, no wrapper is going to fix that.
We need to wrap in a Mutex to get Sync.
We need to wrap the Mutex in an Arc to share it.
We need the type wrapped by Mutex be Sized for reassignment. Fn + Send is ?Sized, so we need to wrap it in a Box to get Sized. Mercifully we can take ownership of these Boxes at function calls so that client code is passing us only the Box rather than the whole Arc<Mutex<Box pile.