factor internal structure of GameImpl. Add set_lock_delay.
A bunch of Arcs shared by GameImpl and GameInfo are now wrapped up in a struct named SharedState. Due to a weirdness in deriving Clone, the obvious Clone implementation had to be done manually. This avoids a bunch of duplications when adding new shared state, and the manual clone is made up for by the simplified clone in run().